pyproject.toml
[tool.poetry]
name = "crm-gbvy-backend"
version = "0.1.0"
description = ""
authors = ["Bruno B. <brunoamix04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
black = "^24.8.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
django-ninja = "^1.3.0"
pydantic = {extras = ["email"], version = "^2.9.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
